完善数据保存和接口

+++ src/main/java/cn/edu/cup/hilly/dataSource/model/mongo/result/ResultAllLineFP.java
@@ -23,5 +23,5 @@ public class ResultAllLineFP {
     private Integer segments;
     private Double singleLength;
     private double[][] aLineFP;
-    private Map<Integer, double[]> allLineFPMap;
+    private Map<Double, double[]> allLineFPMap;
 }
