+++ src/main/java/cn/edu/cup/hilly/dataSource/model/mongo/result/ResultMHis.java
@@ -19,5 +19,5 @@ public class ResultMHis {
     private Integer segments;
     private Double singleLength;
     private double[][] mHis;
-    private Map<Integer, double[]> mHisMap;
+    private Map<Double, double[]> mHisMap;
 }
